move locations
