+++ src/meta/src/stream/mod.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 mod scale;
-mod scheduler;
 mod source_manager;
 mod stream_graph;
 mod stream_manager;
@@ -22,7 +21,6 @@ mod test_fragmenter;
 mod test_scale;
 
 pub use scale::*;
-pub use scheduler::*;
 pub use source_manager::*;
 pub use stream_graph::*;
 pub use stream_manager::*;
